refactor(check): tidy new gate code per self-scan
The #1055 additions tripped the dogfooded gate in three places, and
each had a real simplification rather than a rebaseline:

- dispatch.rs repeated the same optional-counter bump four times;
  a `bump_tally` helper replaces the `if let Some(counter)` blocks
  and returns `validate_and_resolve_file` under its cognitive and
  halstead limits.
- The skipped-input measurement and summary move to a
  `commands/check/skipped.rs` submodule, the same boundary the six
  existing check submodules draw, taking check.rs back under the
  file-level ploc limit.
- The strict-profile flip becomes `CheckArgs::apply_strict`,
  alongside the type's other resolution methods.

`run_check` keeps one genuinely new pipeline step and lands 3% over
its recorded halstead.effort, so the baseline is refreshed with
`make self-scan-write-baseline-headroom` in this same change, per
the baseline-refresh discipline.

Refs #1055
